feat: add only artist route logic without linking to other resources

+++ src/artist/artist.controller.ts
@@ -3,13 +3,18 @@ import {
   Get,
   Post,
   Body,
-  Patch,
   Param,
   Delete,
+  Put,
+  HttpCode,
+  HttpStatus,
 } from '@nestjs/common';
 import { ArtistService } from './artist.service';
 import { CreateArtistDto } from './dto/create-artist.dto';
 import { UpdateArtistDto } from './dto/update-artist.dto';
+import { validate } from 'uuid';
+import { ArtistIDError } from './errors/artist-id-error';
+import { ArtistNotFoundError } from './errors/artist-not-found';
 
 @Controller('artist')
 export class ArtistController {
@@ -27,16 +32,28 @@ export class ArtistController {
 
   @Get(':id')
   findOne(@Param('id') id: string) {
-    return this.artistService.findOne(id);
+    if (!validate(id)) throw new ArtistIDError();
+    const result = this.artistService.findOne(id);
+
+    if (typeof result === 'boolean') throw new ArtistNotFoundError();
+    return result;
   }
 
-  @Patch(':id')
+  @Put(':id')
   update(@Param('id') id: string, @Body() updateArtistDto: UpdateArtistDto) {
-    return this.artistService.update(id, updateArtistDto);
+    if (!validate(id)) throw new ArtistIDError();
+
+    const result = this.artistService.update(id, updateArtistDto);
+    if (typeof result === 'boolean') throw new ArtistNotFoundError();
+    return result;
   }
 
   @Delete(':id')
+  @HttpCode(HttpStatus.NO_CONTENT)
   remove(@Param('id') id: string) {
-    return this.artistService.remove(id);
+    if (!validate(id)) throw new ArtistIDError();
+
+    const result = this.artistService.remove(id);
+    if (typeof result === 'boolean') throw new ArtistNotFoundError();
   }
 }
